Moves hamburger button test into modular one-column feature test. Home page test wasn't rendering page correctly.
