+++ app/views/_cascade/blocks/html/_cascade_assets.html
@@ -1,6 +1,2 @@
 <link href="/_assets/master.css" media="all" rel="stylesheet"/>
 <script src="/_assets/master.js"></script>
-
-<!-- keep these here even if refreshing the above _assets/master* files -->
-<link href="//www.chapman.edu/_files/css/omni-nav.min.css?ver=6-28-16" rel="stylesheet" type="text/css"/>
-<script src="//www.chapman.edu/_files/js/omni-nav.min.js?ver=2-10-16" type="text/javascript"></script>
